+++ src/js/LottieView.js
@@ -7,28 +7,39 @@ import {
   Platform,
   StyleSheet,
   ViewPropTypes,
-  requireNativeComponent
+  requireNativeComponent,
+  NativeModules,
 } from 'react-native';
 import SafeModule from 'react-native-safe-modules';
 import PropTypes from 'prop-types';
 
-const NativeLottieView = requireNativeComponent("LottieAnimationView");
-//  SafeModule.component({
-//   viewName: 'LottieAnimationView',
-//   mockComponent: View,
-// });
+
+const getNativeLottieViewForDesktop = () => {
+  return requireNativeComponent('LottieAnimationView') 
+}
+
+const NativeLottieView =
+  Platform.OS === 'macos' || Platform.OS === 'windows' ?
+    getNativeLottieViewForDesktop() :
+    SafeModule.component({ viewName: 'LottieAnimationView', mockComponent: View })
+
 const AnimatedNativeLottieView = Animated.createAnimatedComponent(NativeLottieView);
 
-const LottieViewManager = SafeModule.module({
-  moduleName: 'LottieAnimationView',
-  mock: {
-    play: () => {},
-    reset: () => {},
-    pause: () => {},
-    resume: () => {},
-    getConstants: () => {},
-  },
-});
+const LottieViewManager = Platform.select({
+  // react-native-windows doesn't work with SafeModule, it always returns the mock component
+  macos: NativeModules.LottieAnimationView,
+  windows: NativeModules.LottieAnimationView,
+  default: SafeModule.module({
+    moduleName: 'LottieAnimationView',
+    mock: {
+      play: () => {},
+      reset: () => {},
+      pause: () => {},
+      resume: () => {},
+      getConstants: () => {},
+    },
+  })
+})
 
 const ViewStyleExceptBorderPropType = (props, propName, componentName, ...rest) => {
   const flattened = StyleSheet.flatten(props[propName] || {});
@@ -68,6 +79,7 @@ const propTypes = {
   onAnimationFinish: PropTypes.func,
   onAnimationLoop: PropTypes.func,
   onLayout: PropTypes.func,
+  cacheComposition: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -77,6 +89,7 @@ const defaultProps = {
   autoPlay: false,
   autoSize: false,
   enableMergePathsAndroidForKitKatAndAbove: false,
+  cacheComposition: true,
   useNativeLooping: false,
   resizeMode: 'contain',
 };
@@ -122,7 +135,7 @@ class LottieView extends React.PureComponent {
   reset() {
     this.runCommand('reset');
   }
-  
+
   pause() {
     this.runCommand('pause');
   }
@@ -144,13 +157,14 @@ class LottieView extends React.PureComponent {
           safeGetViewManagerConfig('LottieAnimationView').Commands[name],
           args,
         ),
-      ios: () => LottieViewManager[name](this.getHandle(), ...args),
       windows: () =>
         UIManager.dispatchViewManagerCommand(
           handle,
           safeGetViewManagerConfig('LottieAnimationView').Commands[name],
           args,
         ),
+      ios: () => LottieViewManager[name](this.getHandle(), ...args),
+      macos: () => LottieViewManager[name](this.getHandle(), ...args),
     })();
   }